refactor(NcGuestContent): migrate component to Typescript
- for https://github.com/nextcloud-libraries/nextcloud-vue/issues/841
