+++ src/components/index.ts
@@ -54,8 +54,8 @@ export { default as NcDialogButton } from './NcDialogButton/index.js'
 export { default as NcEllipsisedOption } from './NcEllipsisedOption/index.js'
 export { default as NcEmojiPicker } from './NcEmojiPicker/index.js'
 export { default as NcEmptyContent } from './NcEmptyContent/index.ts'
-export { default as NcGuestContent } from './NcGuestContent/index.js'
-export { default as NcHeaderButton } from './NcHeaderButton/index'
+export { default as NcGuestContent } from './NcGuestContent/index.ts'
+export { default as NcHeaderButton } from './NcHeaderButton/index.ts'
 export { default as NcHeaderMenu } from './NcHeaderMenu/index.js'
 export { default as NcHighlight } from './NcHighlight/index.js'
 export { default as NcIconSvgWrapper } from './NcIconSvgWrapper/index.js'
